Correct unsigned integer conversion issues
When are reading an unsigned 32 bit value into a long we need to
make sure the top byte is cast to a long before we shift it into
position, otherwise it may make the int value negative and that
would sign extend into a negative 64 bit value, rather than being
a positive 64 bit value in the range 2-4G.

+++ org.spearce.jgit.test/tst/org/spearce/jgit/lib/XInputStream.java
@@ -54,7 +54,7 @@ class XInputStream extends BufferedInputStream {
 
 	long readUInt32() throws IOException {
 		readFully(intbuf, 0, 4);
-		return (intbuf[0] & 0xff) << 24 | (intbuf[1] & 0xff) << 16
+		return ((long) (intbuf[0] & 0xff)) << 24 | (intbuf[1] & 0xff) << 16
 				| (intbuf[2] & 0xff) << 8 | (intbuf[3] & 0xff);
 	}
 }
